mxshop/goods_srv/requirements.txt
peewee
pymysql
passlib
grpcio
grpcio-tools
loguru
protobuf
nacos-sdk-python
python-consul

